requirements-dev.txt
# Kubeflow RAG Template - Development Requirements
# Install with: pip install -r requirements-dev.txt

# Core dependencies
kfp>=2.0.0
kfp-kubernetes>=1.1.0
pyyaml>=6.0

# Ingestion
pypdf>=3.0.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Embedding & Vector Stores
openai>=1.12.0
anthropic>=0.20.0
sentence-transformers>=2.6.0
chromadb>=0.4.22

# Generation
tiktoken>=0.6.0

# Evaluation
ragas>=0.1.7
datasets>=2.18.0

# API Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Dev & Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
ruff>=0.3.0
black>=24.0.0
mypy>=1.9.0
httpx>=0.27.0

# Utilities
tqdm>=4.66.0
numpy>=1.26.0
pandas>=2.2.0
